Moved: DdsFilter to Commands Main Module
