+++ projects/tools/dxt-lossless-transform-cli/src/main.rs
@@ -7,35 +7,8 @@ mod debug;
 mod error;
 mod util;
 use argh::FromArgs;
+use commands::DdsFilter;
 use core::error::Error;
-use std::str::FromStr;
-
-#[derive(Debug, Clone)]
-pub enum DdsFilter {
-    BC1,
-    BC2,
-    BC3,
-    BC7,
-    All,
-}
-
-// Implement FromStr to allow parsing from command line arguments
-impl FromStr for DdsFilter {
-    type Err = String;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s.to_lowercase().as_str() {
-            "bc1" => Ok(DdsFilter::BC1),
-            "bc2" => Ok(DdsFilter::BC2),
-            "bc3" => Ok(DdsFilter::BC3),
-            "bc7" => Ok(DdsFilter::BC7),
-            "all" => Ok(DdsFilter::All),
-            _ => Err(format!(
-                "Invalid DDS type: {s}. Valid types are: bc1, bc2, bc3, bc7, all"
-            )),
-        }
-    }
-}
 
 #[derive(FromArgs, Debug)]
 /// File transformation tool for DDS files
